Change metadata action to get any tag for docker tags
Previously it was only allowing tags that were valid semantic versioning https://semver.org and ignoring them otherwise. This meant versions with and underscore _ instead of a dash were not successfully tagging the images. Now any tag should work.

It is still recommended to use valid semantic versioning for tags.